MR-5   Cast one of the operands of this multiplication operation to a "long". src/jmemorize/core/Card.java
src/jmemorize/util/TimeSpan.java

+++ src/jmemorize/util/TimeSpan.java
@@ -30,7 +30,7 @@ import jmemorize.gui.Localization;
  */
 public class TimeSpan
 {
-    public final static long MINUTE = 60 * 1000;
+    public final static long MINUTE = (long) 60 * 1000;
     public final static long HOUR   = 60 * MINUTE;
     public final static long DAY    = 24 * HOUR;
